move users API out of NCR namespace
