+++ config/routes.rb
@@ -8,9 +8,10 @@ C2::Application.routes.draw do
   namespace :api, constraints: lambda {|req| ENV['API_ENABLED'] == 'true' } do
     scope :v1 do
       namespace :ncr do
-        resources :users, only: [:index]
         resources :work_orders, only: [:index]
       end
+
+      resources :users, only: [:index]
     end
   end
